move dbconfig to golem-common module
