+++ golem-component-service/src/config.rs
@@ -15,12 +15,13 @@
 use serde::{Deserialize, Serialize};
 use std::path::Path;
 
-use golem_common::config::{ConfigExample, ConfigLoader, HasConfigExamples};
+use golem_common::config::{
+    ConfigExample, ConfigLoader, DbConfig, DbSqliteConfig, HasConfigExamples,
+};
 use golem_common::tracing::TracingConfig;
 use golem_component_service_base::config::ComponentCompilationConfig;
 use golem_service_base::config::{
-    ComponentStoreConfig, ComponentStoreLocalConfig, ComponentStoreS3Config, DbConfig,
-    DbSqliteConfig,
+    ComponentStoreConfig, ComponentStoreLocalConfig, ComponentStoreS3Config,
 };
 use golem_service_base::model::Empty;
